Q2Pt2
The S, I, and R variables are able to determine the outcomes of an SIR model. During an infection simulation, if the starting number of individuals who are infected is high, the probability of susceptible person becoming infected is also high. However, as more people become infected, this probability decreases as the number of susceptible people decreases. At the end of the simulation, the number of recovered individuals is the number of people affected by the disease transmission in the population. This number can ultimately be affected by two variables, beta and gamma. Beta is the value which relates to the rate of transmission for the disease. If this number is high, it will lead to an increased number of infected individuals as well as resistant individuals provided the disease is not lethal. Gamma is the value which accounts for the rate at which an infected individual recovers and becomes a resistant individual. If gamma values are low, the rate at which resistance occurs is low, meaning there are more opportunities for an individual to infect susceptible people. This means that the number of infected individuals will increase as gamma decreases. The SIR model is dependent upon these five variables, but it is also dependent upon time. The disease landscape will change as time passes.